requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.9.1
    # via django
boto3==1.40.11
    # via -r requirements.in
botocore==1.40.11
    # via
    #   boto3
    #   s3transfer
django==5.2.5
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown==3.8.2
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
urllib3==2.5.0
    # via botocore
